feat: Add S3Controller functions for interacting with AWS S3 bucket
The code changes in `s3Controller.js` add several functions to interact with an AWS S3 bucket. These functions include getting the contents of a folder, uploading a file, creating a new folder, deleting a file, getting the sort_order tag data from a file, and setting the sort_order tag data for a file.

This commit message follows the established convention of using a prefix to indicate the type of change (`feat` for a new feature) and provides a concise summary of the changes made.