feat: Add ClientDB class for interacting with IndexedDB
The code changes in `clientDB.js` introduce a new class `ClientDB` that provides methods for initializing the IndexedDB, saving client data, and retrieving client data. This class will be used to interact with the `clientDataStore` object store in the IndexedDB.

This commit message follows the established convention of using a prefix to indicate the type of change (`feat` for a new feature) and provides a concise summary of the changes made.